requirements.txt
# Core dependencies for AI Data Matching Tool
streamlit>=1.37.0
pandas>=2.3.2
rapidfuzz>=3.13.0
plotly>=5.24.0
openpyxl>=3.1.3
jinja2>=3.1.6
requests>=2.32.5

# OpenAI for LLM capabilities
openai>=1.50.0

# Additional utilities
python-dotenv>=1.0.1
xlrd>=2.0.1